refactor(release): add browser graph, help popups, and code cleanup
- Add interactive browser-based dependency graph with Cytoscape.js
  - Dagre/Force/Circle layouts, search, zoom, PNG export
  - CLI flags: --web to open browser, --out to save HTML
  - 'g' key in TUI opens browser graph

- Add help popups (h key) to release TUIs
  - release graph: navigation, browser graph, quit shortcuts
  - release status: panel switching, scrolling, navigation

- Fix all clippy warnings
  - commit_analyzer: combine if branches, use if-let, implement FromStr
  - config: use derive(Default)
  - session: remove needless borrow
  - version/changelog: remove unnecessary casts

- Remove unused UI components (production cleanup)
  - Deleted: confirm_dialog, panel, popup, toast
  - Simplified utils.rs to only is_interactive_terminal()
  - Each TUI has its own implementation, no shared abstraction needed

+++ src/core/release/changelog.rs
@@ -8,7 +8,6 @@
 //! has to do with looking at the repository history since the most recent
 //! release(s). That's exactly the information contained in a release changelog.
 
-use time::OffsetDateTime;
 use std::{
     collections::HashMap,
     fs::File,
@@ -16,6 +15,7 @@ use std::{
     path::PathBuf,
 };
 use thiserror::Error as ThisError;
+use time::OffsetDateTime;
 
 use super::dynfmt::{Format, SimpleCurlyFormat};
 
@@ -279,7 +279,12 @@ impl Changelog for MarkdownChangelog {
         let now = OffsetDateTime::now_utc();
         header_args.insert(
             "yyyy_mm_dd",
-            format!("{:04}-{:02}-{:02}", now.year(), now.month() as u8, now.day() as u8),
+            format!(
+                "{:04}-{:02}-{:02}",
+                now.year(),
+                now.month() as u8,
+                now.day()
+            ),
         );
 
         let changelog_path = self.changelog_path(proj, repo);
